Config

+++ erysasystem/src/main/java/com/erysa/system/erysasystem/repositorio/ProductoRepositorio.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.erysa.system.erysasystem.modelo.Producto;
 
 @Repository
-public interface ProductoRepositorio extends PagingAndSortingRepository<Producto, Long> {
+public interface ProductoRepositorio extends PagingAndSortingRepository<Producto, Integer> {
 
 }
